Refactor so that samples are in their own namespace
