+++ rules/rules.go
@@ -1,0 +1,6 @@
+package rules
+
+import "embed"
+
+//go:embed *
+var FS embed.FS
