+++ samples/Linux/2021.ua-parser-js/preinstall.js
