+++ samples/Linux/2022.Magneto/magnet_goblin_dropper.sh
