+++ samples/Linux/2023.FreeDownloadManager/freedownloadmanager_clear_postinst
