+++ samples/Linux/2024.Mirai_Nomi/mirai.crontab
