+++ samples/Linux/clean/ping.x86_64
