+++ samples/Linux/clean/redis-server.aarch64
