+++ samples/does-nothing/does-nothing.simple
