+++ samples/macOS/2024.SpectralBlur.DPRK/SpectralBlur-macshare.md
@@ -1,4 +1,4 @@
-## testdata/macOS/SpectralBlur-macshare
+## macOS/2024.SpectralBlur.DPRK/SpectralBlur-macshare
 
 Overall risk: 🔥 3/HIGH
 
@@ -32,4 +32,3 @@ Overall risk: 🔥 3/HIGH
 | 1/LOW    | process/multithreaded   | uses pthreads                                       |
 | 1/LOW    | process/username/get    | get login name                                      |
 | 1/LOW    | random/insecure         | generate random numbers insecurely                  |
-
